### 2025-04-17 #8072 @editorialbot
[PRE REVIEW]: CFTime.jl: a Julia package for representing time following the Climate and Forecast conventions
**Submitting author:** @Alexander-Barth (<a href="http://orcid.org/0000-0003-2952-5997">Alexander Barth</a>)
**Repository:** https://github.com/JuliaGeo/CFTime.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.1
