+++ paper/paper.bib
@@ -39,9 +39,6 @@
   commit = {90ed5641684604096558a77020038583e1f2459f}
 }
 
-Eaton, B., Gregory, J., Drach, B., Taylor, K., Hankin, S. et al. (2024). NetCDF Climate and Forecast (CF) Metadata Conventions (1.12). CF Community. https://doi.org/10.5281/zenodo.14275599
-
-
 @misc{Eaton2024,
   author       = {Eaton, Brian and
                   Gregory, Jonathan and
