+++ paper/paper.bib
@@ -4,7 +4,8 @@
   year          =  2022,
   month         =  6,
   number        = "21-050r1",
-  url          = "https://www.opengis.net/doc/CS/zarr/2.0"
+  url          = "https://www.opengis.net/doc/CS/zarr/2.0",
+  doi = "10.62973/21-050r1"
 }
 
 @article{Barth2022,
@@ -333,5 +334,6 @@
   year = {2025},
   note = {R package version 1.6.0},
   url = {https://CRAN.R-project.org/package=CFtime},
-  note = {(last access: 4 April 2025)}
+  note = {(last access: 4 April 2025)},
+  doi = {10.32614/cran.package.cftime}
 }
