+++ paper/paper.bib
@@ -30,15 +30,6 @@
   url = {https://doi.org/10.21105/joss.02018}
 }
 
-@misc{NCDatasets,
-  author = {Alexander Barth},
-  title = {NCDatasets: A julia package for manipulating netCDF data sets},
-  year = {2023},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  howpublished = {\url{https://github.com/Alexander-Barth/NCDatasets.jl}},
-  commit = {90ed5641684604096558a77020038583e1f2459f}
-}
 
 @misc{Eaton2024,
   author       = {Eaton, Brian and
