### 2025-04-23 6d50f5b
shorten paper and move some info to documentation
