+++ paper/paper.md
@@ -23,17 +23,16 @@ bibliography: paper.bib
 
 
 Climate and Forecast (CF) conventions are a metadata standard for Earth data [@Eaton2024] and are mainly used in oceanography and meteorology.
-They aim to be equally applicable to model data, remote-sensing data and in-situ data, despite the high heterogeneity among data types.
 The CF conventions were originally proposed for the NetCDF storage format, but they are also increasingly used with other formats like Zarr [@OGC_Zarr] and GRIB ([GRIBDatasets](https://github.com/JuliaGeo/GRIBDatasets.jl)).
 
-Since the initial release of the Climate and Forecast (CF) conventions [@Eaton2003], the encoding of time instances has been standardized. This part of the standard is based on the earlier COARDS (Cooperative Ocean Atmosphere Research Data Service) standard from 1995 [@COARDS]. The Julia package CFTime implements various standardized calendars that have been standardized in the frame of these conventions. It also supports some arithmetic operations for example, computing the duration between two time instances, ordering time instances and creating time ranges. The time origin and resolution are flexible ranging from days to attoseconds.
+Since the initial release of the Climate and Forecast (CF) conventions [@Eaton2003], the encoding of time instances has been standardized. The Julia package CFTime implements various calendars that have been standardized in the frame of these conventions. It also supports some arithmetic operations for example, computing the duration between two time instances, ordering time instances and creating time ranges. The time origin and resolution are flexible ranging from days to attoseconds.
 
 
 # Statement of need
 
 In many Earth science disciplines and beyond, expressing a time instance and a duration is essential. The CF conventions provide a rich and flexible
 framework for handling time, equally applicable to observations and model data. To our knowledge, CFTime.jl is the only package in the Julia ecosystem that implements the time structures standardized by the CF conventions. While almost all datasets used in Earth Science use dates after the year 1582, some datasets or software systems use a time origin before this date, which makes it necessary to handle the transition from Julian to Gregorian calendar [@Octave; @SeaDataNet_format].
-Some users also expressed the need for microseconds and nanoseconds as time resolution. In particular, the popular Python package Pandas [@mckinney-proc-scipy-2010] defaults to nanoseconds as the time resolution. It was therefore necessary to also support sub-second time resolutions, even if they are rarely used in typical Earth science applications.
+Some users also expressed the need for microseconds and nanoseconds as time resolution even if they are rarely used in typical Earth science applications.
 
 As of 31 March 2025, 119 Julia packages depend directly or indirectly on CFTime (excluding optional dependencies). CFTime is for example used by numerical models, such as ClimaOcean.jl, a framework for realistic ocean and coupled sea-ice simulations based on Oceananigans.jl [@OceananigansJOSS], the hydrological modeling package Wflow.jl [@vanVerseveld2024] and AIBECS.jl, a modeling framework for global marine biogeochemical cycles [@Pasquier2022].
 
@@ -59,14 +58,11 @@ The CF conventions define several calendar types, including:
 | Calendar                | Type                         | Explanation |
 | ----------------------- | ---------------------------- | ---------------------------- |
 | `standard`, `gregorian` | `DateTimeStandard`           | the Gregorian calendar after 15 October 1582 and the Julian calendar before  |
-| `proleptic_gregorian`   | `DateTimeProlepticGregorian` | the Gregorian calendar applied to all dates (including before 15 October 1582) |
[…]
