### 2025-04-23 #8072 @Alexander-Barth
Hi @kthyng ! Thank you for taking a look at the paper.
I added the two DOI suggestions:
https://github.com/JuliaGeo/CFTime.jl/commit/226f9340697e5f026204f8b6a4f2fab42c496fff

I also removed the unused citation to the github repository of NCDatasets.jl. I use the previous JOSS paper of NCDatasets.jl (with DOI) in the references:
https://github.com/JuliaGeo/CFTime.jl/commit/c28cbf12b98818a729a9ffa11c9acd14d49a95ac

Unfortunately, I am not aware of a DOI for the other links.

I would suggest the following reviewers (some experts on the CF conventions and some earth-science data analysis in julia) :

sadielbartholomew
davidhassell
felixcremer
natgeo-wong
milankl

(github handle with the leading @ removed).
I can confirm that I do not have a conflict of interest with these potential reviewers.

I also shortened the paper. The update paper has been reduced to 947 words.
Let me know if you have any other comments.
Thanks again!
