### 2025-06-05 #8072 @haakon-e
Hi @mengqi-z, I'm happy to review this submission. I will likely be able to start looking at this by the end of next week.

I don't think this classifies as a COI, but to be sure would like to flag that I've previously interacted a fair bit with one of Alexander's other packages ([link to issue tracker](https://github.com/JuliaGeo/NCDatasets.jl/issues?q=author%3Ahaakon-e)).
