### 2025-06-09 #8072 @eviatarbach
Thank you for the invitation! Unfortunately, I do not have time at the moment.
