### 2025-06-18 #8072 @mengqi-z
👋 @OmarShehata @kellykochanski - Would any of you be available to review this submission for JOSS? JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
