### 2025-06-18 JuliaGeo/CFTime.jl#33 @haakon-e
Describe how to test the package
Could you add a short note somewhere that describes how to test the package?

I eventually figured out that I simply do `] test` from the root environment, but in my experience, packages approach this in a few different ways. So, it may be helpful to simply state briefly how to run tests.

ref: https://github.com/openjournals/joss-reviews/issues/8487
