+++ test/test_time.jl
@@ -2,6 +2,10 @@ using CFTime
 import Dates
 using Dates: DateTime, Day, @dateformat_str
 using Test
+import CFTime: datetuple
+datetuple(dt::DateTime) = (Dates.year(dt),Dates.month(dt),Dates.day(dt),
+                            Dates.hour(dt),Dates.minute(dt),Dates.second(dt),
+                            Dates.millisecond(dt))
 
 # slow, but accurate and easy to understand (and possibly fix)
 include("reference_algorithm.jl")
@@ -29,65 +33,71 @@ function datenum_datetuple_all_calendars(::Type{T}) where T
     end
 end
 
-for T in [DateTimeStandard, DateTimeJulian, DateTimeProlepticGregorian,
+@testset "datenum_datetuple_all_calendars" begin
+    for T in [DateTimeStandard, DateTimeJulian, DateTimeProlepticGregorian,
           DateTimeAllLeap, DateTimeNoLeap, DateTime360Day]
-    datenum_datetuple_all_calendars(T)
+        datenum_datetuple_all_calendars(T)
+    end
 end
+
 # test of DateTime structures
+@testset "DateTime structures" begin
+    dt = DateTimeNoLeap(1959,12,31,23,39,59,123)
+    @test dt + Dates.Millisecond(7) == DateTimeNoLeap(1959,12,31,23,39,59,130)
+    @test dt + Dates.Second(7)      == DateTimeNoLeap(1959,12,31,23,40,6,123)
+    @test dt + Dates.Minute(7)      == DateTimeNoLeap(1959,12,31,23,46,59,123)
+    @test dt + Dates.Hour(7)        == DateTimeNoLeap(1960,1,1,6,39,59,123)
+    @test dt + Dates.Day(7)         == DateTimeNoLeap(1960,1,7,23,39,59,123)
+    @test dt + Dates.Month(7)       == DateTimeNoLeap(1960,7,31,23,39,59,123)
+    @test dt + Dates.Year(7)        == DateTimeNoLeap(1966,12,31,23,39,59,123)
+    @test dt + Dates.Month(24)      == DateTimeNoLeap(1961,12,31,23,39,59,123)
 
-dt = DateTimeNoLeap(1959,12,31,23,39,59,123)
-@test dt + Dates.Millisecond(7) == DateTimeNoLeap(1959,12,31,23,39,59,130)
-@test dt + Dates.Second(7)      == DateTimeNoLeap(1959,12,31,23,40,6,123)
-@test dt + Dates.Minute(7)      == DateTimeNoLeap(1959,12,31,23,46,59,123)
-@test dt + Dates.Hour(7)        == DateTimeNoLeap(1960,1,1,6,39,59,123)
-@test dt + Dates.Day(7)         == DateTimeNoLeap(1960,1,7,23,39,59,123)
-@test dt + Dates.Month(7)       == DateTimeNoLeap(1960,7,31,23,39,59,123)
-@test dt + Dates.Year(7)        == DateTimeNoLeap(1966,12,31,23,39,59,123)
-@test dt + Dates.Month(24)      == DateTimeNoLeap(1961,12,31,23,39,59,123)
+    @test dt - Dates.Month(0)       == DateTimeNoLeap(1959,12,31,23,39,59,123)
+    @test dt - Dates.Month(24)      == DateTimeNoLeap(1957,12,31,23,39,59,123)
+    @test dt - Dates.Year(7)        == DateTimeNoLeap(1952,12,31,23,39,59,123)
+end
 
-@test dt - Dates.Month(0)       == DateTimeNoLeap(1959,12,31,23,39,59,123)
[…]
