### 2025-06-18 JuliaGeo/CFTime.jl#37 @haakon-e
Add formatter
To aid developers and enforce a consistent style within codebase, I would recommend adding a formatter of your choice to the repo, e.g.
- [JuliaFormatter.jl](https://github.com/domluna/JuliaFormatter.jl): allows additional customization
- [Runic.jl](https://github.com/fredrikekre/Runic.jl): does not allow any customization

This can be added as a `.github` workflow for convenient checks in PRs.

Personally, I like `JuliaFormatter.jl` with these options:
```
# .JuliaFormatter.toml
margin = 120
indent = 4
always_for_in = true
whitespace_typedefs = true
whitespace_ops_in_indices = true
join_lines_based_on_source = true  # remove_extra_newlines/#join_lines_based_on_source
```
in particular, the high `margin=120` + `join_lines_based_on_source=true` allow for a bit of flexibility to the local context.

----

This is a low-priority addition, but one that I think would be convenient as more developers start contributing to the package.

If you do this, information about applying the formatter should be added to the developer docs.

ref: https://github.com/openjournals/joss-reviews/issues/8487
