### 2025-06-18 JuliaGeo/CFTime.jl#38 @haakon-e
performance of `T(10)^(-n)` for `n <= 0`
With reference to:
https://github.com/JuliaGeo/CFTime.jl/blob/6693cc13e13c4032d1817081ddb68144ccddc55d/src/period.jl#L76-L88
I don't have access to an NVIDIA GPU right now, but have you tried `exp10`?

In some very small and quick tests on my Mac,
- on the Metal GPU, I get identical timings using `T(10)^(-n)` vs `exp10(-n)`
- on the CPU, I find that `exp10` is ~40% faster.

ref: https://github.com/openjournals/joss-reviews/issues/8487
