### 2025-06-19 4dcac9e
Information about how to test the package (issue #33)
