### 2025-06-19 b0db2b4
Information about how to test the package (issue #33)
