### 2025-06-19 JuliaGeo/CFTime.jl#33 @Alexander-Barth
Good idea, I have added this:

https://github.com/JuliaGeo/CFTime.jl/commit/b0db2b4ab51430dbf2ac885cefc3f36eb85c7159

(I did not mention the root environment in the added instruction, as the tests can be run in any environment where CFTime is installed)
