+++ test/test_time.jl
@@ -10,14 +10,23 @@ datetuple(dt::DateTime) = (Dates.year(dt),Dates.month(dt),Dates.day(dt),
 # slow, but accurate and easy to understand (and possibly fix)
 include("reference_algorithm.jl")
 
-# reference value from Meeus, Jean (1998)
-# launch of Sputnik 1
+@testset "Check with reference values" begin
+    # reference value from Meeus, Jean (1998)
+    # launch of Sputnik 1
+    @test CFTime.datetuple_ymd(DateTimeStandard,2_436_116 - 2_400_001) == (1957, 10, 4)
+    @test CFTime.datenum_gregjulian(1957,10,4,true) == 36115
+    @test CFTime.datenum_gregjulian(333,1,27,false) == -557288
 
-@test CFTime.datetuple_ymd(DateTimeStandard,2_436_116 - 2_400_001) == (1957, 10, 4)
-@test CFTime.datenum_gregjulian(1957,10,4,true) == 36115
 
-@test CFTime.datenum_gregjulian(333,1,27,false) == -557288
-
+    # reference values from python's cftime
+    @test DateTimeJulian(2000,1,1) + Dates.Day(1) == DateTimeJulian(2000,01,02)
+    @test DateTimeJulian(2000,1,1) + Dates.Day(12) == DateTimeJulian(2000,01,13)
+    @test DateTimeJulian(2000,1,1) + Dates.Day(123) == DateTimeJulian(2000,05,03)
+    @test DateTimeJulian(2000,1,1) + Dates.Day(1234) == DateTimeJulian(2003,05,19)
+    @test DateTimeJulian(2000,1,1) + Dates.Day(12345) == DateTimeJulian(2033,10,19)
+    @test DateTimeJulian(2000,1,1) + Dates.Day(12346) == DateTimeJulian(2033,10,20)
+    @test DateTimeJulian(1,1,1) + Dates.Day(1234678) == DateTimeJulian(3381,05,14)
+end
 
 function datenum_datetuple_all_calendars(::Type{T}) where T
     #dayincrement = 11
@@ -33,28 +42,11 @@ function datenum_datetuple_all_calendars(::Type{T}) where T
     end
 end
 
-@testset "datenum_datetuple_all_calendars" begin
+@testset "Internal consistency" begin
     for T in [DateTimeStandard, DateTimeJulian, DateTimeProlepticGregorian,
           DateTimeAllLeap, DateTimeNoLeap, DateTime360Day]
         datenum_datetuple_all_calendars(T)
     end
-end
-
-# test of DateTime structures
-@testset "DateTime structures" begin
-    dt = DateTimeNoLeap(1959,12,31,23,39,59,123)
-    @test dt + Dates.Millisecond(7) == DateTimeNoLeap(1959,12,31,23,39,59,130)
-    @test dt + Dates.Second(7)      == DateTimeNoLeap(1959,12,31,23,40,6,123)
-    @test dt + Dates.Minute(7)      == DateTimeNoLeap(1959,12,31,23,46,59,123)
-    @test dt + Dates.Hour(7)        == DateTimeNoLeap(1960,1,1,6,39,59,123)
-    @test dt + Dates.Day(7)         == DateTimeNoLeap(1960,1,7,23,39,59,123)
-    @test dt + Dates.Month(7)       == DateTimeNoLeap(1960,7,31,23,39,59,123)
-    @test dt + Dates.Year(7)        == DateTimeNoLeap(1966,12,31,23,39,59,123)
-    @test dt + Dates.Month(24)      == DateTimeNoLeap(1961,12,31,23,39,59,123)
-
[…]
