+++ test/runtests.jl
@@ -2,18 +2,34 @@ using CFTime
 using Test
 
 @testset verbose = true "All tests" begin
-    @testset verbose = true "Time and calendars" begin
+    @testset verbose = true "Basic functionality" begin
         include("test_time.jl")
     end
-    @testset "Resolution" begin
+
+    @testset "Operators (+,-,<,>...)" begin
+        include("test_operators.jl")
+    end
+
+    @testset verbose = true "CF conventions" begin
+        include("test_cf.jl")
+    end
+
+    @testset "Different time resolutions" begin
         include("test_resolution.jl")
     end
+
     @testset "Rounding" begin
         include("test_rounding.jl")
     end
+
     @testset "Year 0" begin
         include("test_year0.jl")
     end
+
+    @testset "Reported issues" begin
+        include("test_issues.jl")
+    end
+
     @testset "Aqua" begin
         include("test_aqua.jl")
     end
