### 2025-06-19 d80c0ed
refactor tests, add test_cf.jl test_operators.jl test_rounding.jl  test_issues.jl
