+++ test/test_cf.jl
@@ -1,0 +1,171 @@
+using CFTime
+
+
+# time units
+@testset "CF time units" begin
+    t0,plength = CFTime.timeunits("days since 1950-01-02T03:04:05Z")
+    @test t0 == DateTimeStandard(1950,1,2, 3,4,5)
+    @test plength == 86400000
+
+
+    t0,plength = CFTime.timeunits("days since -4713-01-01T00:00:00Z")
+    @test t0 == DateTimeStandard(-4713,1,1)
+    @test plength == 86400000
+
+
+    t0,plength = CFTime.timeunits("days since -4713-01-01")
+    @test t0 == DateTimeStandard(-4713,1,1)
+    @test plength == 86400000
+
+
+    t0,plength = CFTime.timeunits("days since 2000-01-01 0:0:0")
+    @test t0 == DateTimeStandard(2000,1,1)
+    @test plength == 86400000
+
+    t0,plength = CFTime.timeunits("days since 2000-01-01 00:00")
+    @test t0 == DateTimeStandard(2000,1,1)
+    @test plength == 86400000
+
+    # issue 24
+    t0,plength = CFTime.timeunits("hours since 1900-01-01 00:00:00.0")
+    @test t0 == DateTimeStandard(1900,1,1)
+    @test plength == 86400000 ÷ 24
+
+
+    t0,plength = CFTime.timeunits("seconds since 1992-10-8 15:15:42.5")
+    @test t0 == DateTimeStandard(1992,10,8,15,15,42,500)
+    @test plength == 1000
+
+    units = "microseconds since 2000-01-01T23:59:59.12345678"
+    origintuple, ratio = timeunits(Tuple,units)
+    @test origintuple == (2000, 1, 1, 23, 59, 59, 123, 456, 780)
+
+
+    for (calendar,DT) in [
+        ("standard",DateTimeStandard),
+        ("gregorian",DateTimeStandard),
+        ("proleptic_gregorian",DateTimeProlepticGregorian),
+        ("julian",DateTimeJulian),
+        ("noleap",DateTimeNoLeap),
+        ("365_day",DateTimeNoLeap),
+        ("all_leap",DateTimeAllLeap),
+        ("366_day",DateTimeAllLeap),
+        ("360_day",DateTime360Day)]
+
+        calendart0,calendarplength = CFTime.timeunits("days since 2000-1-1 0:0:0",calendar)
+        @test calendart0 == DT(2000,1,1)
+        @test calendarplength == 86400000
+    end
+
+    @test_throws ErrorException CFTime.timeunits("fortnights since 2000-01-01")
+    @test_throws ErrorException CFTime.timeunits("days since 2000-1-1 0:0:0","foo")
+end
+
+@testset "Decoding/decoding CF time data" begin
+
+    # value from python's cftime
+    # print(cftime.DatetimeJulian(-4713,1,1) + datetime.timedelta(2455512,.375 * 24*60*60))
+    # 2010-10-29 09:00:00
+
+    #@test timedecode([2455512.375],"days since -4713-01-01T00:00:00","julian")
+    #   == DateTimeJulian(2010,10,29,09)
+
+    # values from
+    # https://web.archive.org/web/20180212214229/https://en.wikipedia.org/wiki/Julian_day
+
+    # Modified JD
+    @test CFTime.timedecode([58160.6875],"days since 1858-11-17","standard" )==
+        [DateTime(2018,2,11,16,30,0)]
+
+    # CNES JD
+    @test CFTime.timedecode([24878.6875],"days since 1950-01-01","standard") ==
+        [DateTime(2018,2,11,16,30,0)]
+
[…]
