### 2025-06-19 bcf3817
add test with reference algorithm (issue #34)
