### 2025-06-19 1100780
add test with reference algorithm (issue #34)
