### 2025-06-19 114beb4
add test with reference algorithm (issue #34)
