### 2025-06-19 500e61d
group tests using the reference algorithm (issue #34)
