### 2025-06-19 2a47d26
update time units tests (issue #34)
