+++ test/test_issues.jl
@@ -1,0 +1,184 @@
+using CFTime
+
+# issue #12
+
+units = "days since 1850-01-01 00:00:00"
+calendar = "noleap"
+data_orig = [54750.5, 54751.5, 54752.5]
+
+# Decoding
+datacal = CFTime.timedecode(data_orig, units, calendar)
+# Reencoding
+data_orig_back = CFTime.timeencode(datacal, units, calendar)
+@test data_orig ≈ data_orig_back
+
+
+# issue #17
+
+# reference values from cftime
+# for T in [cftime.DatetimeGregorian,cftime.DatetimeJulian,cftime.DatetimeProlepticGregorian,cftime.DatetimeAllLeap,cftime.DatetimeNoLeap, cftime.Datetime360Day]:
+#     print(T,T(1582,11,1) - T(1582,10,1))
+
+@test daysinmonth(DateTimeStandard(1582,10,1)) == 21
+@test daysinmonth(DateTimeJulian(1582,10,1)) == 31
+@test daysinmonth(DateTimeProlepticGregorian(1582,10,1)) == 31
+@test daysinmonth(DateTimeAllLeap(1582,10,1)) == 31
+@test daysinmonth(DateTimeNoLeap(1582,10,1)) == 31
+@test daysinmonth(DateTime360Day(1582,10,1)) == 30
+
+# import cftime
+# for T in [cftime.DatetimeGregorian,cftime.DatetimeJulian,cftime.DatetimeProlepticGregorian,cftime.DatetimeAllLeap,cftime.DatetimeNoLeap, cftime.Datetime360Day]:
+#    print(T,T(1583,1,1) - T(1582,1,1))
+
+@test daysinyear(DateTimeStandard(1582,10,1)) == 355
+@test daysinyear(DateTimeJulian(1582,10,1)) == 365
+@test daysinyear(DateTimeProlepticGregorian(1582,10,1)) == 365
+@test daysinyear(DateTimeAllLeap(1582,10,1)) == 366
+@test daysinyear(DateTimeNoLeap(1582,10,1)) == 365
+@test daysinyear(DateTime360Day(1582,10,1)) == 360
+
+
+for T in [DateTimeStandard, DateTimeJulian, DateTimeProlepticGregorian,
+          DateTimeAllLeap, DateTimeNoLeap, DateTime360Day]
+    @test Dates.yearmonthday(T(2004,1,2)) == (2004, 1, 2)
+    @test Dates.yearmonth(T(2004,1,2)) == (2004, 1)
+    @test Dates.monthday(T(2004,1,2)) == (1, 2)
+
+    # test constructor with argument which are not Int64
+    @test T(Int16(2000),Int32(1),UInt8(1)) == T(2000,1,1)
+end
+
+# time ranges
+
+@test length(DateTimeNoLeap(2000, 01, 01):Dates.Day(1):DateTimeNoLeap(2000, 12, 31)) == 365
+@test length(DateTimeNoLeap(2000, 01, 01):Dates.Month(1):DateTimeNoLeap(2000, 12, 31)) == 12
+
+for T in [DateTimeStandard, DateTimeJulian, DateTimeProlepticGregorian,
+          DateTimeAllLeap, DateTimeNoLeap, DateTime360Day]
+
+    # end date is inclusive
+    @test length(T(2000, 01, 01):Dates.Month(1):T(2001, 1, 1)) == 13
+    @test length(T(2000, 01, 01):Dates.Year(1):T(2001, 1, 1)) == 2
+end
+
+# issue #21
+@test parse(DateTimeNoLeap,"1999-12-05", dateformat"yyyy-mm-dd") == DateTimeNoLeap(1999,12,05)
[…]
