### 2025-06-23 2a70ad5
delete test_time_units.jl as tests are now in test_cf.jl (issue #34)
