### 2025-06-23 42c442d
remove unfinished GPU test (issue #34)
