### 2025-06-23 b8c4571
avoid horizontal scoll-bars in docs (issue #36)
