+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"#CFTime.jl","page":"CFTime.jl","title":"CFTime.jl","text":"","category":"section"},{"location":"","page":"CFTime.jl","title":"CFTime.jl","text":"This package implements the calendar types from the CF convention, namely:","category":"page"},{"location":"","page":"CFTime.jl","title":"CFTime.jl","text":"Mixed Gregorian/Julian calendar  (DateTimeStandard)\nProleptic gregorian calendar (DateTimeProlepticGregorian)\nGregorian calendar without leap years (all years are 365 days long) (DateTimeNoLeap)\nGregorian calendar with only leap year (all years are 366 days long) (DateTimeAllLeap)\nA calendar with every year being 360 days long (divided into 30 day months) (DateTime360Day)\nJulian calendar (DateTimeJulian)","category":"page"},{"location":"","page":"CFTime.jl","title":"CFTime.jl","text":"Note that time zones and leap seconds are currently not supported by CFTime.jl.","category":"page"},{"location":"#Installation","page":"CFTime.jl","title":"Installation","text":"","category":"section"},{"location":"","page":"CFTime.jl","title":"CFTime.jl","text":"Inside the Julia shell, you can download and install the package by issuing:","category":"page"},{"location":"","page":"CFTime.jl","title":"CFTime.jl","text":"using Pkg\nPkg.add(\"CFTime\")","category":"page"},{"location":"#Latest-development-version","page":"CFTime.jl","title":"Latest development version","text":"","category":"section"},{"location":"","page":"CFTime.jl","title":"CFTime.jl","text":"If you want to try the latest development version, you can do this with the following commands:","category":"page"},{"location":"","page":"CFTime.jl","title":"CFTime.jl","text":"using Pkg\nPkg.add(PackageSpec(url=\"https://github.com/JuliaGeo/CFTime.jl\", rev=\"master\"))","category":"page"},{"location":"#Types","page":"CFTime.jl","title":"Types","text":"","category":"section"},{"location":"#CFTime.DateTimeStandard","page":"CFTime.jl","title":"CFTime.DateTimeStandard","text":"DateTimeStandard([Ti::DataType], y, [m, d, h, mi, s, ms, ...], origin = (1900, 1, 1), units = ...) -> DateTimeStandard\n\nConstruct a DateTimeStandard type by year (y), month (m, default 1), day (d, default 1), hour (h, default 0), minute (mi, default 0), second (s, default 0), millisecond (ms, default 0), .... Currently attosecond is the smallest supported time unit.\n\nAll arguments must be convertible to Int64. DateTimeStandard is a subtype of AbstractCFDateTime.\n\nThe date is stored a duration since the time origin (epoch) expressed as milliseconds but smaller time units will be used if necessary. For example if the user provides 8 integers, they will be interpreted as year, month, day, hour, minute, second, millisecond and microsecond. The internal time units will be microsecond in this case.\n\nThe netCDF CF calendars are defined in the CF Standard. This type implements the calendar defined as \"standard\".\n\n\n\n\n\nDateTimeStandard(dt::AbstractString, format::AbstractString; locale=\"english\") -> DateTimeStandard\n\nConstruct a DateTimeStandard by parsing the dt date time string following the pattern given in the format string.\n\nnote: Note\nThis function is experimental and might be removed in the future. It relies on some internal function
[…]
