### 2025-06-23 JuliaGeo/CFTime.jl#38 @Alexander-Barth
In order to maintain precision, all these numbers need to be integers for type `T`. If we use `exp10`, we get a floating point number: `exp10(3) = 1000.0`. CFTime can also work with `Int128` and `BigInt`. We would lose this feature if somewhere computations are done using floating points numbers.

For type-stable code, this function is actually evaluated during compilation (thanks to inlining and constant propagation) so there is no runtime overhead. But for some reason, the GPU compiler has problems with this unless I use the commented function `ten_tom` which works on GPU and CPU but makes the code slower on CPU. This is basically the reasons why the GPU support is not ready and I did not mention it in the JOSS paper.

In any case, I did not know about `exp10` ! If not here, it will come handy somewhere else.
