### 2025-06-23 #8072 @mengqi-z
👋 @alextbradley @MartinuzziFrancesco - Would any of you be willing to review this submission for JOSS? JOSS follows a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
