### 2025-06-27 e9b3d65
capture type T using ::Type{T}
