### 2025-07-15 #8487 @mengqi-z
👋 @haakon-e, @alextbradley, @MartinuzziFrancesco - Just checking in to see how the review is going. Could each of you please provide a brief status update in this thread? If you anticipate any delays, please let me know. Thanks!
