### 2025-07-22 JuliaGeo/CFTime.jl#39 @alextbradley
Documentation tweaks
Could you add a clearer description of the statement of need for this package in the documentation? I think this is quite nicely explained in the paper and you could probably cut and paste across. (Though, coming from a field where CF convention usage/discussion is less common, it did take me a while to figure out the gap that this package fills; perhaps you could add a higher level, layman summary of the problems the software is designed to solve?).

Similarly, it would be good to state explicitly in the installation steps in the documentation that there are no dependencies required ("CFTime is a pure Julia package and currently depends only on the modules Dates and Printf, which are part of Julia’s standard library.")

ref https://github.com/openjournals/joss-reviews/issues/8487
