### 2025-07-22 JuliaGeo/CFTime.jl#40 @alextbradley
Documentation/examples
I am currently working through the documentation. Some notes below (apologies if these should be obvious to me, I am coming at this as a non-expert in the specific topic!)

- For each of the types, it would be good to add explicit examples. You should also add explanation of the arguments in the constructor functions (e.g. what does ```[Ti::DataType]``` refer to?)
- It would be good to explicitly state that the types are the core 'calendar' objects that the package interacts with
- In the timedecode function, state what the format of the ```units``` string has to be
- Add an example for timeencode
- I know they are just wrapper functions for the Dates package, but why not just use CFTime for all of the accessor function documentations, rather than Dates for those down to nanosecond and CFTime for smaller?

In addition, it would be good to make the example usages more explicit (I'd suggest adding an 'examples' section to the documentation) and add some examples where the package is used to solve real world issues (at the moment the issues are quite abstract).

ref https://github.com/openjournals/joss-reviews/issues/8487
