### 2025-07-22 47d9641
statement of need in documentation (issue #39)
