### 2025-07-22 JuliaGeo/CFTime.jl#39 @Alexander-Barth
Thank you, @alextbradley, for your suggestion and I agree with it! I revised the documentation in [47d9641](https://github.com/JuliaGeo/CFTime.jl/commit/47d964159c50af071ac2e83cad691f8306fe1b87).
