### 2025-07-23 e0cd534
more examples and document the parameters Ti, units and locale (issue #40)
