+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"#CFTime.jl","page":"CFTime.jl","title":"CFTime.jl","text":"","category":"section"},{"location":"","page":"CFTime.jl","title":"CFTime.jl","text":"In many Earth science disciplines and beyond, expressing a time instance and a duration is essential. The CF conventions provide a rich and flexible framework for handling time, equally applicable to observations and model data.","category":"page"},{"location":"","page":"CFTime.jl","title":"CFTime.jl","text":"CFTime.jl implements the time structures standardized by the CF conventions, namely:","category":"page"},{"location":"","page":"CFTime.jl","title":"CFTime.jl","text":"Julian calendar (DateTimeJulian). A year is a leap year if it is divisible by 4. For example, 1900 and 2000 are both leap years in the Julian calendar.\nProleptic Gregorian calendar (DateTimeProlepticGregorian). A year is a leap year if it is divisible by 4 but not 100 or if it is divisible by 400. For example, 1900 is not a leap year in the proleptic Gregorian calendar but 2000 is.\nMixed Gregorian/Julian calendar  (DateTimeStandard). This calendar uses the Julian calendar for time instances before 15th October 1582 and Gregorian calendar afterwards.\nA calendar without leap years (DateTimeNoLeap). All years are 365 days long.\nA calendar with only leap years (DateTimeAllLeap). All years are 366 days long.\nA calendar with every year being 360 days long (divided into 30-day months) (DateTime360Day).","category":"page"},{"location":"","page":"CFTime.jl","title":"CFTime.jl","text":"The first three calendars (with different rules for leap years) can be used to express the time instances of observations or model. The remaining three calendars correspond to idealised model configurations where the duration of a year (revolution of the Earth around the Sun) is assumed to be exactly 365, 366 or 360 days.","category":"page"},{"location":"","page":"CFTime.jl","title":"CFTime.jl","text":"While almost all datasets used in Earth Science use dates after the year 1582, some datasets or software systems use a time origin before this date, which makes it necessary to handle the transition from Julian to Gregorian calendar. Additionally, some dataset use microseconds and nanoseconds as time resolution, whereas Julia's Dates.DateTime has milliseconds as time resolution.","category":"page"},{"location":"","page":"CFTime.jl","title":"CFTime.jl","text":"Note that time zones and leap seconds are currently not supported by CFTime.jl.","category":"page"},{"location":"#Installation","page":"CFTime.jl","title":"Installation","text":"","category":"section"},{"location":"","page":"CFTime.jl","title":"CFTime.jl","text":"Inside the Julia shell, you can download and install the package by issuing:","category":"page"},{"location":"","page":"CFTime.jl","title":"CFTime.jl","text":"using Pkg\nPkg.add(\"CFTime\")","category":"page"},{"location":"","page":"CFTime.jl","title":"CFTime.jl","text":"CFTime is a pure Julia package and currently depends only on the modules Dates and Printf, which are part of Julia’s standard library.","category":"page"},{"location":"#Latest-development-version","page":"CFTime.jl","title":"Latest development version","text":"","category":"section"},{"location":"","page":"CFTime.jl","title":"CFTime.jl","text":"If you want to try the latest development version, you can do this with the following commands:","ca
[…]
