### 2025-07-24 c7dc29e
implement floor(CFTime.Period,...)
