### 2025-07-24 76918f9
add example_CMIP6.jl
