### 2025-07-24 b55b95f
add version of dependencies for documentation
