+++ docs/Project.toml
@@ -1,9 +1,20 @@
 [deps]
+CFTime = "179af706-886a-5703-950a-314cd64e0468"
 CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
+CommonDataModel = "1fbeeb36-5f17-413c-809b-666fb144f157"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
 Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
+NCDatasets = "85f8d34a-cbdd-5861-8df4-14fed0d494ab"
+Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
 [compat]
 Documenter = "1"
 Literate = "2"
+CommonDataModel = "0.3"
+Documenter = "1.14"
+NCDatasets = "0.14"
+Statistics = "1"
+Dates = "1"
+Downloads = "1"
