### 2025-07-24 ac5afb3
add plots
