### 2025-07-24 f756a13
fix docs/Project.toml
