+++ docs/Project.toml
@@ -13,7 +13,6 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Documenter = "1"
 Literate = "2"
 CommonDataModel = "0.3"
-Documenter = "1.14"
 NCDatasets = "0.14"
 Statistics = "1"
 Dates = "1"
