### 2025-07-24 bcea28d
tweak example
