### 2025-07-25 3b09f18
role of core types (issue #40)
