### 2025-07-25 63dda7b
mention that DateTimeStandard is the default calendar
