### 2025-07-25 bbad9db
fix typos in example
