### 2025-07-25 0fa62fb
improve wording in example
