### 2025-07-25 8a42974
more generic floor
