### 2025-07-28 JuliaGeo/CFTime.jl#42 @haakon-e
Mixed time type arithmetic
The following result surprises me as a user:
```Julia
julia> dtnl = DateTimeNoLeap(2015,1,1,12)
DateTimeNoLeap(2015-01-01T12:00:00)

julia> dts = DateTimeStandard(1582,10,4)
DateTimeStandard(1582-10-04T00:00:00)

julia> dts + (dtnl - dts)
DateTimeStandard(3871-07-17T12:00:00)
```

It may be useful to either:
- restrict certain operations on mixed types, or
- provide detailed documentation / warnings on this type of behavior.

The core issue, of course, is that:
```Julia
julia> DateTimeNoLeap(2015,1,1,12) - DateTimeNoLeap(1582,10,4)
13631284800000 milliseconds

julia> DateTimeNoLeap(2015,1,1,12) - DateTimeStandard(1582,10,4)
72226209600000 milliseconds
```

---

ref: https://github.com/openjournals/joss-reviews/issues/8487
