### 2025-07-28 JuliaGeo/CFTime.jl#42 @haakon-e
Here's a related, surprising comparison example:
```julia
julia> DateTimeStandard(1960,9,27) > DateTimeNoLeap(1900,1,1)
false
```
