### 2025-07-28 JuliaGeo/CFTime.jl#43 @haakon-e
Constructor allows inconsistent inputs
For example, setting the `hour` value different from 0, but declaring units in `:day`

```julia
julia> t1 = DateTimeStandard(2001,1,1,12; units=:day)
DateTimeStandard(2001-01-01T00:00:00)

julia> t2 = DateTimeStandard(2001,1,1,0; units=:day)
DateTimeStandard(2001-01-01T00:00:00)

julia> t1 == t2
true

julia> t1 === t2
true

```

---

ref: https://github.com/openjournals/joss-reviews/issues/8487
