### 2025-07-28 JuliaGeo/CFTime.jl#44 @haakon-e
Discrepancy in printing of type compared to docs
In the docs for [timedecode](https://juliageo.org/CFTime.jl/dev/#CFTime.timedecode), the example suggests printing shows in a certain format:

```julia
dt = CFTime.timedecode([0,1,2,3],"days since 2000-01-01 00:00:00","360_day")
# 4-element Array{DateTime360Day,1}:
#  DateTime360Day(2000-01-01T00:00:00)
#  DateTime360Day(2000-01-02T00:00:00)
#  DateTime360Day(2000-01-03T00:00:00)
#  DateTime360Day(2000-01-04T00:00:00)
```

If I type the same in REPL, it instead shows as
```julia

julia> dt = CFTime.timedecode([0,1,2,3],"days since 2000-01-01 00:00:00","360_day")
4-element Vector{DateTime360Day{CFTime.Period{Int64, Val{86400}(), Val{0}()}, Val{(2000, 1, 1)}()}}:
 DateTime360Day(2000-01-01T00:00:00)

 DateTime360Day(2000-01-02T00:00:00)

 DateTime360Day(2000-01-03T00:00:00)

 DateTime360Day(2000-01-04T00:00:00)

```

related to #41. This issue, in particular, is concerned with the additional type information that is shown.

---

ref: https://github.com/openjournals/joss-reviews/issues/8487
