### 2025-07-28 JuliaGeo/CFTime.jl#45 @haakon-e
Documented output type for `yearmonthday`
The documented output type for [`yearmonthday`](https://juliageo.org/CFTime.jl/dev/#Dates.yearmonthday) is:
```Julia
"""
    yearmonthday(dt::AbstractCFDateTime) -> (Int64, Int64, Int64)

Simultaneously return the year, month and day parts of `dt`.
"""
```

However, this is not always correct. Consider the following example:
```Julia-repl
julia> t2 = DateTimeStandard(Int128,2001,1,1)
DateTimeStandard(2001-01-01T00:00:00)

julia> yearmonthday(t2)
(2001, 1, 1)

julia> yearmonthday(t2) |> typeof
Tuple{Int128, Int128, Int128}

```

this also applies to related functions.

----

ref: https://github.com/openjournals/joss-reviews/issues/8487
