### 2025-07-28 JuliaGeo/CFTime.jl#46 @haakon-e
Add docstring for `AbstractCFDateTime`
It would be useful with a docstring for `AbstractCFDateTime` so that any mention of the symbol in the docs can be cross-referenced, for example: ``[`AbstractCFDateTime`](@ref)``.

Currently, the help page is auto-generated. Specifically, it would be useful with a user-friendly summary and explanations of the various subtypes of this abstract type.

```Julia-repl
help?> AbstractCFDateTime
search: AbstractCFDateTime AbstractMatrix AbstractUnitRange AbstractPattern

  No documentation found for public symbol.

  Summary
  ≡≡≡≡≡≡≡

  abstract type AbstractCFDateTime{T, Torigintuple}

  Subtypes
  ≡≡≡≡≡≡≡≡

  CFTime.DateTimeTAI{T, Torigintuple}
  CFTime.DateTimeUTC{T, Torigintuple}
  DateTime360Day{T, Torigintuple}
  DateTimeAllLeap{T, Torigintuple}
  DateTimeJulian{T, Torigintuple}
  DateTimeNoLeap{T, Torigintuple}
  DateTimeProlepticGregorian{T, Torigintuple}
  DateTimeStandard{T, Torigintuple}

  Supertype Hierarchy
  ≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡≡

  AbstractCFDateTime{T, Torigintuple} <: TimeType <: Dates.AbstractTime <: Any
```

---

ref: https://github.com/openjournals/joss-reviews/issues/8487
