### 2025-07-28 JuliaGeo/CFTime.jl#47 @haakon-e
Add methods for `Picosecond(::Integer)` and below
You can do e.g. `Second(1)` to get a `Dates` "`TimePeriod`"-like object of said duration. `Dates` support periods down to `Nanosecond`. For completeness, it would be useful if this package added support for `Picosecond`, `Femtosecond`, `Attosecond`.

I see that there is support for such short durations through e.g. `Period(1, :picosecond)`. This constructor is not documented, so I presume it's internal and should not be used, although the `struct` itself _is_ documented. Clarification on this would be useful.

---

That said, due to the numerous references to the `Dates` package, as well as extensive interoperability, it is somewhat confusing that both this and that package defines the name `Period`. Moreover, in `Dates`, it is an `abstract type`, whilst here it is a `struct` (though not `export`ed, presumably to avoid name conflicts). To avoid confusion, would it be better to consider another name for the object here, e.g. `CFPeriod`?

---

ref: https://github.com/openjournals/joss-reviews/issues/8487
