### 2025-07-28 JuliaGeo/CFTime.jl#48 @haakon-e
A few edge cases during `CFDateTime` construction
# "All arguments must be convertible to `Int64`"
See docstring, e.g., for [DateTimeStandard](https://juliageo.org/CFTime.jl/dev/#CFTime.DateTimeStandard).

Intuitively, I should be able to supply a `float` for which `convert(Int, 1.0)` succeeds. This is not the case.
```Julia-repl
julia> DateTimeStandard(1)
DateTimeStandard(0001-01-01T00:00:00)

julia> DateTimeStandard(1.0)
ERROR: MethodError: no method matching DateTimeStandard(::Float64)
The type `DateTimeStandard` exists, but no method is defined for this combination of argument types when trying to construct it.
[...]

julia> convert(Int, 1.0)
1
```

# Usage of `Int128`

## The need to explicitly set integer type

If you're not careful, you can run into issues by adjusting the time units without adjusting the internal storage type. I think it would be useful if the second call below either errored or emitted a clear warning:

```Julia-repl
julia> DateTimeStandard(2000,12,31, units = :nanosecond)
DateTimeStandard(2000-12-31T00:00:00)

julia> DateTimeStandard(2000,12,31, units = :picosecond)
DateTimeStandard(1858-12-14T12:46:30.088578957312)

julia> DateTimeStandard(Int128, 2000,12,31, units = :picosecond)
DateTimeStandard(2000-12-31T00:00:00)
```

---

With the following integers (which should be interpreted as "years" in the constructors, I believe):

```Julia-repl
julia> i_64 = typemax(Int64)
9223372036854775807

julia> i_128 = Int128(i_64)
9223372036854775807

julia> ip1_128 = i_128 + 1
9223372036854775808
```

## a big time, with `:attosecond` units

```Julia-repl
julia> DateTimeStandard(i_64; units = :attosecond)
DateTimeStandard(1858-11-16T-1:00:00.261099644983967744)
```
As you can see, this produces a reasonable-looking date, except that the hour is `-1`.

Unfortunately, explicitly setting the storage type does not help:
```Julia-repl
julia> DateTimeStandard(Int128, i_64; units = :attosecond)
DateTimeStandard(-1101089977897-10-09T20:15:39.409414765026476032)

# same with all inputs as `Int128`
julia> DateTimeStandard(Int128, i_128; units = :attosecond)
DateTimeStandard(-1101089977897-10-09T20:15:39.409414765026476032)

# changing the units gives a different result
julia> DateTimeStandard(Int128, i_128; units = :picosecond)
DateTimeStandard(-1559714879324475638-05-10T18:09:52.568231788544)
```

A likely related issue arises from adding a small time unit to an otherwise reasonable time:
```Julia-repl
julia> t_ok = DateTimeStandard(2001,1,1,12)
DateTimeStandard(2001-01-01T12:00:00)

julia> Δt = CFTime.Period(1, :attosecond)
1 attosecond

julia> t_bad = t_ok + Δt
DateTimeStandard(1858-11-16T-1:00:00.165100303580921857)
```
(ref: #47, which ponders whether this way to construct `1 attosecond` is "allowed")

probing the internals, it's clear that `typemax(Int64)` was exceeded:
```Julia
julia> t_bad.instant
4693045119639617537 attoseconds
```
for this particular instance `t_ok`, this also happens for `1 picosecond`. I wonder if it would be more helpful to either do automatic type promotion, or error.

---

ref: https://github.com/openjournals/joss-reviews/issues/8487
