### 2025-07-28 JuliaGeo/CFTime.jl#49 @haakon-e
Document use cases for `Int128`
Following up to the `Int128` issues in #48, I think it would be useful if the docs did a more thorough explanation of the utility of `Int128` (and beyond). This includes:
- "common" instances when `Int128` may be useful[^1], and/or
- heuristics for when to use `Int128`. For example by considering the range of representable times for a given unit.
- performance trade-offs.
- type stability with `Int128`
- etc...

[^1]: I suspect this is actually not commonly needed at all.

---

ref: https://github.com/openjournals/joss-reviews/issues/8487
