### 2025-07-28 JuliaGeo/CFTime.jl#46 @haakon-e
If you're open to reconsidering names of various objects (ref: #47), I would suggest naming this object `CFDateTime` instead of `AbstractCFDateTime`. It's more concise, and all concrete implementations (`DateTimeJulian`, etc.) have names that are suffixed to `DateTime` anyways.

This follows precedence, for example with the `abstract type Period` from [`Dates.jl`](https://github.com/JuliaLang/julia/blob/55b1f3c41baeb6b0ccfc0ad89d7c941e6c8bc229/stdlib/Dates/src/types.jl#L21).
