### 2025-07-28 73128e2
docs: add xrefs to docs
