### 2025-07-28 JuliaGeo/CFTime.jl#51 @haakon-e
paper: state of the field
With regards to

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

it seems most obvious to me that a comparison is made to the built-in `Dates` library in Julia. Notably
- which calendar type is/are supported by `Dates`
- how does `Dates` conform to, and deviate from, CF conventions
- with the additional features `CFTime` offers, including multiple calendar types and sub-microsecond resolution, would it be feasible to provide some concrete and relevant examples (or brief mention) from the large number (119!) of Julia packages that need these features?
- Relatedly, it would be nice to have a citation providing an example for the "need for microseconds [...] as time resolution", if possible.

---

ref: https://github.com/openjournals/joss-reviews/issues/8487
