### 2025-07-28 ffb2982
info about why using @eval here
