### 2025-07-28 876da5e
update example output (issue #44)
