+++ src/conversions.jl
@@ -333,18 +333,18 @@ not possible for the other calendars.
 using CFTime, Dates
 # standard calendar
 dt = CFTime.timedecode([0,1,2,3],"days since 2000-01-01 00:00:00")
-# 4-element Array{Dates.DateTime,1}:
+# 4-element Vector{Dates.DateTime}:
 #  2000-01-01T00:00:00
 #  2000-01-02T00:00:00
 #  2000-01-03T00:00:00
 #  2000-01-04T00:00:00
 
 dt = CFTime.timedecode([0,1,2,3],"days since 2000-01-01 00:00:00","360_day")
-# 4-element Array{DateTime360Day,1}:
-#  DateTime360Day(2000-01-01T00:00:00)
-#  DateTime360Day(2000-01-02T00:00:00)
-#  DateTime360Day(2000-01-03T00:00:00)
-#  DateTime360Day(2000-01-04T00:00:00)
+# 4-element Vector{DateTime360Day{CFTime.Period{Int64, Val{86400}(), Val{0}()}, Val{(2000, 1, 1)}()}}:
+#  2000-01-01T00:00:00
+#  2000-01-02T00:00:00
+#  2000-01-03T00:00:00
+#  2000-01-04T00:00:00
 ```
 
 """
