### 2025-07-28 JuliaGeo/CFTime.jl#44 @Alexander-Barth
Thanks a lot, this has been addressed in [876da5e](https://github.com/JuliaGeo/CFTime.jl/commit/876da5e1b67c6b444bfc9664892a2ba2c03622cc).
