### 2025-07-28 cc79527
issue #52
