### 2025-07-28 JuliaGeo/CFTime.jl#47 @Alexander-Barth
I think there is a misunderstanding.
Concerning the public API, I follow the approach of Julia:
https://docs.julialang.org/en/v1/manual/faq/#How-does-Julia-define-its-public-API?

In particular:
> Functions, types, and constants are not part of the public API if they are not public, even if they have docstrings [...].

`CFTime.Period` is not public API for this package (even if it has a docstring) as the type is also not described in the documentation.

In generally, I try to minimize the public API to avoid breaking changes. Uses are quite annoyed and sometimes vocal about it, despite a package not have reached 1.0.
