### 2025-07-28 74c55da
fast comparision if possible
