### 2025-07-28 5fc5109
raise an error on some mixed time type arithmetic (issue #42)
