### 2025-07-28 JuliaGeo/CFTime.jl#42 @Alexander-Barth
Thank you for reporting that issue. It has been fixed in [5fc5109](https://github.com/JuliaGeo/CFTime.jl/commit/5fc51098574c3e061d20eb739fd67f19acc19e91).
Now, a method error is raised for differences and comparisons that do not make sense.

Can you close the issue, unless you think that the issue is not solved entirely? An issue can also be reopened later if that should be necessary.
