### 2025-07-28 JuliaGeo/CFTime.jl#47 @haakon-e
Thank you for the comment.

Let me approach this another way. The [docs](https://juliageo.org/CFTime.jl/dev/#Ranges) demonstrate how to construct ranges for the `CFTime` types:

> Time ranges can be constructed using a start date, end date and a time increment:
>
> ```
> range = DateTimeStandard(2000,1,1):Dates.Day(1):DateTimeStandard(2000,12,31)
> length(range)
> # returns 366
> step(range)
> # returns 1 day
> ```
>
> Note that there is no default increment for range.

I think it is reasonable for a user to be under the impression that `CFTime` can construct ranges for all supported "time units". So, how can I construct a range with picosecond (and below) increments?

If the answer is equivalent to: "this is not supported and I do not plan to support this", then I would ask that this is transparently explained in the docs.
