### 2025-07-28 JuliaGeo/CFTime.jl#44 @haakon-e
Thanks!
