### 2025-07-28 #8487 @haakon-e
A note for everyone: I am mostly done with the review. I will continue to engage with Alexander on the issues I've opened through this week. Next week I will be unavailable, and then available again the week of August 11.

Alexander: Thanks for your thoughtful comments and code improvements so far in the process!
