### 2025-07-29 JuliaGeo/CFTime.jl#53 @MartinuzziFrancesco
Make use of `@example`s in the documentation
I am running the documentation examples, and they are quite easy to follow. However, the blocks are run in `julia` mode which provides julia syntax highlighting. A good addiction would be to wrap them instead in `@example`, in order to make them run every time the docs are compiled. This solution would help catch some small errors like the following:

in [Rounding](https://juliageo.org/CFTime.jl/stable/#Rounding) we have:

```julia
using CFTime: DateTimeStandard
using Dates: DateTime
dt = DateTimeStandard
```
which works as advertised, however the next block

```julia
using CFTime: DateTimeStandard

dt = DateTimeStandard(24*60*60,"second since 2000-01-01")

floor(dt+Second(9),Second(10)) == dt
```

returns

```julia
ERROR: UndefVarError: `Second` not defined in `Main`
Suggestion: check for spelling errors or missing imports.
Hint: a global variable of this name may be made accessible by importing Dates in the current active module Main
Stacktrace:
 [1] top-level scope
   @ REPL[8]:1
```

Of course this is not really an issue, but having `@example` blocks would remove this tiny errors and assure reproducibility of the docs

ref https://github.com/openjournals/joss-reviews/issues/8487
