### 2025-07-29 014ba4b
use jldoctest (issue #53)
