### 2025-07-29 JuliaGeo/CFTime.jl#53 @Alexander-Barth
Thank you for this nice suggestion. This issue has been addressed in the two commits. I forgot to repeat "using Dates". I converted multiple code blocks to `jldoctest` (similar to @example) in order to test also for the output value.
