+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"example_CMIP6/#Example","page":"Example","title":"Example","text":"","category":"section"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"This example shows how to use the CFTime package with CMIP6 data in the NetCDF format following the CF Conventions.","category":"page"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"The CMIP6 example data is from NOAA-GFDL and data extracted at a single point as provided by the R package CFtime. The data is downloaded automatically.","category":"page"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"This example can also be downloaded as a Julia source file example_CMIP6.jl.","category":"page"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"using NCDatasets, CFTime, Dates, Statistics\nusing Downloads: download","category":"page"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"Download the sample data and open the NetCDF file","category":"page"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"url = \"https://github.com/R-CF/CFtime/raw/1509a2387a92bda8500d1d6ac472b36df3575b56/inst/extdata/pr_day_GFDL-ESM4_ssp245_r1i1p1f1_gr1_20150101-20991231_v20180701.nc\"\nfname = download(url);\nds = NCDataset(fname);","category":"page"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"Get the global attribute Conventions","category":"page"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"ds.attrib[\"Conventions\"]","category":"page"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"\"CF-1.7 CMIP-6.0 UGRID-1.0\"","category":"page"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"Get the calendar and units attribute of the variable time","category":"page"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"calendar = ds[\"time\"].attrib[\"calendar\"]\nunits = ds[\"time\"].attrib[\"units\"]\n(calendar,units)","category":"page"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"(\"noleap\", \"days since 1850-01-01\")","category":"page"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"Load the raw data representing the number of days since 1850-01-01 with the no-leap calendar and decode the data. The variable time is a vector of DateTimeNoLeap","category":"page"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"data = ds[\"time\"].var[:];\ntime = CFTime.timedecode(data,units,calendar);","category":"page"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"Since CFTime is integrated in NCDatasets, this transformation above is handled automatically by using:","category":"page"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"time = ds[\"time\"][:];","category":"page"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"Load the precipitation which is a variable of size 1 x 1 x 31025","category":"page"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"pr = ds[\"pr\"][1,1,:];\npr_units = ds[\"pr\"].attrib[\"units\"];\nclose(ds);","category":"page"},{"location":"example_CMIP6/","page":"Example","title":"Example","text":"Check that the time resolution of the datasets is 1 day (and constant).","category":"page"},{"location":"example_C
[…]
