### 2025-08-01 73a65fd
give examples in the documentation about overflows (issue #48)
