### 2025-08-01 5b99eb0
contructor and some operators for Picosecond(::Integer) and below (issue #47)
