### 2025-08-05 JuliaGeo/CFTime.jl#51 @alextbradley
Agree with all of the points @haakon-e makes above! It would also be worth mentioning whether this is a pure Julia implementation of the cftime python package (totally fine!) or whether there are additional features
