### 2025-08-05 JuliaGeo/CFTime.jl#54 @alextbradley
Add community guidelines
Please add community guidelines: describing how to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support
