### 2025-08-05 #8487 @alextbradley
I’ve now finished working through the checklist — congrats on a great job with the code and improvements, Alexander. There are several outstanding items on the checklist, which should be easy to fix and I am happy to recommend publication once completed:

* State of the field: see issue https://github.com/JuliaGeo/CFTime.jl/issues/51.
* Example usage: I tested out the example (c.f. https://github.com/JuliaGeo/CFTime.jl/issues/40) and it looks good. I think the inclusion of a real world example is a big improvement. I think you could also add basic usage examples to this section (perhaps in a subfolder) to enable the user to find example usage more easily.
* Community guidelines: please add community guidelines on contributions (https://github.com/JuliaGeo/CFTime.jl/issues/54)

I will be away from this Friday until the end of the month. Happy for the editors to move forwards with acceptance once the above are addressed.
