### 2025-08-11 JuliaGeo/CFTime.jl#48 @haakon-e
Thanks @Alexander-Barth! I really appreciate the depth of the documentation explaining the overflow issues. I will close this as completed.
