### 2025-08-20 06b690b
return error rather than truncating (issue #43)
