### 2025-08-20 a779d40
remove output types which are sometimes misleading (issue #45)
