+++ docs/src/index.md
@@ -78,6 +78,7 @@ CFTime.attosecond(dt::AbstractCFDateTime)
 
 ## Query Functions
 
+
 ```@docs
 daysinmonth
 daysinyear
