### 2025-08-23 2b24f19
Add community guidelines (issue #54)
