### 2025-08-25 2badf36
add basic_usage.jl (issue #40)
