+++ examples/example_CMIP6.jl
@@ -1,4 +1,4 @@
-# # Example
+# # Loading climate date
 #
 # This example shows how to use the CFTime package with CMIP6 data
 # in the NetCDF format following the CF Conventions.
