### 2025-08-25 78cdfa0
add State of the field (issue #51)
