### 2025-08-25 0dd49b3
add docstring for AbstractCFDateTime (issue #46)
