### 2025-08-25 dd50349
use cases for Int128 (issue #49)
