### 2025-08-25 fb6d7b4
divisions between time periods (issue #55)
