### 2025-08-26 JuliaGeo/CFTime.jl#40 @Alexander-Barth
An additional basic usage example has been added.
