### 2025-08-26 JuliaGeo/CFTime.jl#49 @Alexander-Barth
I am closing the issue. Feel free to re-open if you consider the issue unresolved.
