### 2025-08-26 d6a60da
allowing conversion of sub-millisecond DateTimeStandard to DateTime (issue #56)
