+++ test/test_resolution.jl
@@ -2,27 +2,13 @@ using CFTime
 import CFTime:
     AbstractCFDateTime,
     DateTimeStandard,
-    Period,
     datenum,
     datetuple,
     datetuple_ymd,
     parseDT,
     timetuplefrac,
     timeunits
-import Dates
-import Dates:
-    @dateformat_str,
-    DateTime,
-    day,
-    hour,
-    microsecond,
-    millisecond,
-    minute,
-    month,
-    nanosecond,
-    second,
-    value,
-    year
+using Dates
 using Test
 
 function same_tuple(t1,t2)
@@ -32,29 +18,29 @@ function same_tuple(t1,t2)
         all(==(0),t2[len+1:end])
 end
 
-@test timetuplefrac(Period((2*24*60*60  + 3*60*60 + 4*60  + 5)*1000,1))[1:4] == (2,3,4,5)
-@test timetuplefrac(Period((2*24*60*60  + 3*60*60 + 4*60  + 5),1000))[1:4] == (2,3,4,5)
+@test timetuplefrac(CFTime.Period((2*24*60*60  + 3*60*60 + 4*60  + 5)*1000,1))[1:4] == (2,3,4,5)
+@test timetuplefrac(CFTime.Period((2*24*60*60  + 3*60*60 + 4*60  + 5),1000))[1:4] == (2,3,4,5)
 
-@testset "Base methods on Period" begin
-    p = Period(10, 1)
+@testset "Base methods on CFTime.Period" begin
+    p = CFTime.Period(10, 1)
     @test abs(-p) == p
-    @test zero(p) == Period(0, 1)
-    @test one(p) == Period(1, 1)
+    @test zero(p) == CFTime.Period(0, 1)
+    @test one(p) == CFTime.Period(1, 1)
 end
 
 tuf = (2,3,4,5,6,7,8) # day, hour, minute, seconds, ...
 factor = 1
 exponent = -9
-p = Period(tuf,factor,exponent)
+p = CFTime.Period(tuf,factor,exponent)
 @test timetuplefrac(p)[1:length(tuf)] == tuf
 
 @test CFTime._type(p) == CFTime._type(typeof(p))
 
-@test one(p) == Period(1,factor,exponent)
+@test one(p) == CFTime.Period(1,factor,exponent)
 # test promotion rules
 
-p1 = Period(1,:second)
-p2 = Period(1000,:millisecond)
+p1 = CFTime.Period(1,:second)
+p2 = CFTime.Period(1000,:millisecond)
 
 @test promote_type(typeof(p1),typeof(p2)) == typeof(p2)
 
@@ -75,7 +61,7 @@ pp1,pp2 = promote(p1,Dates.Hour(1))
 
 # missing
 
-p = Period(1,:second)
+p = CFTime.Period(1,:second)
 @test ismissing(p == missing)
 @test ismissing(missing == p)
 @test ismissing(DateTimeStandard(2000,1,1) + missing)
@@ -83,17 +69,17 @@ p = Period(1,:second)
 
 # arithmetic
 
-p1 = Period(1,:microsecond)
-p2 = Period(10,:microsecond)
-@test p1+p2 == Period(11,:microsecond)
+p1 = CFTime.Period(1,:microsecond)
+p2 = CFTime.Period(10,:microsecond)
+@test p1+p2 == CFTime.Period(11,:microsecond)
 
 
-p1 = Period(1,:microsecond)
-p2 = Period(10,:nanosecond)
-@test p1+p2 == Period(1010,:nanosecond)
+p1 = CFTime.Period(1,:microsecond)
+p2 = CFTime.Period(10,:nanosecond)
[…]
