### 2025-08-26 dbbd573
fix typos in doc
