+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"basic_usage/#Basic-Usage","page":"Basic Usage","title":"Basic Usage","text":"","category":"section"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"This basic example shows how to create a DateTime structure following the calendar from the CF conventions.","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"using CFTime\nusing Dates","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"A DateTime structure following the standard calendar from the CF conventions can be instantiated by providing the year, month, day, hour, minute, and second:","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"dt0 = DateTimeStandard(1999, 12, 31, 23, 59, 59)","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"DateTimeStandard(1999-12-31T23:59:59)","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"A DateTimeStandard structure can also be created from a string by providing the format. See Dates.DateFormat for details about format specifiers.","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"dt0 = DateTimeStandard(\"1999-12-31T23:59:59\", dateformat\"yyyy-mm-ddTHH:MM:SS\")","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"DateTimeStandard(1999-12-31T23:59:59)","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"A DateTime structure can be converted to a string explicitly using the string function, which follows the ISO 8601 format:","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"string(dt0)","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"\"1999-12-31T23:59:59\"","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"The conversion can also be done implicitly using string interpolation:","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"\"$dt0\"","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"\"1999-12-31T23:59:59\"","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"Different string formats are supported using the Dates.format function:","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"Dates.format(dt0, \"yyyymmdd-HHMMSS\")","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"\"19991231-235959\"","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"Year, month, day, etc. can be extracted from the DateTime structure using the corresponding functions:","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","text":"year   = Dates.year(dt0);\nmonth  = Dates.month(dt0);\nday    = Dates.day(dt0);\nhour   = Dates.hour(dt0);\nminute = Dates.minute(dt0);\nsecond = Dates.second(dt0);","category":"page"},{"location":"basic_usage/","page":"Basic Usage","title":"Basic Usage","tex
[…]
