+++ test/test_resolution.jl
@@ -11,15 +11,15 @@ import CFTime:
 using Dates
 using Test
 
-function same_tuple(t1,t2)
-    len = min(length(t1),length(t2))
-    (t1[1:len] == t2[1:len]) &&
-        all(==(0),t1[len+1:end]) &&
-        all(==(0),t2[len+1:end])
+function same_tuple(t1, t2)
+    len = min(length(t1), length(t2))
+    return (t1[1:len] == t2[1:len]) &&
+        all(==(0), t1[(len + 1):end]) &&
+        all(==(0), t2[(len + 1):end])
 end
 
-@test timetuplefrac(CFTime.Period((2*24*60*60  + 3*60*60 + 4*60  + 5)*1000,1))[1:4] == (2,3,4,5)
-@test timetuplefrac(CFTime.Period((2*24*60*60  + 3*60*60 + 4*60  + 5),1000))[1:4] == (2,3,4,5)
+@test timetuplefrac(CFTime.Period((2 * 24 * 60 * 60 + 3 * 60 * 60 + 4 * 60 + 5) * 1000, 1))[1:4] == (2, 3, 4, 5)
+@test timetuplefrac(CFTime.Period((2 * 24 * 60 * 60 + 3 * 60 * 60 + 4 * 60 + 5), 1000))[1:4] == (2, 3, 4, 5)
 
 @testset "Base methods on CFTime.Period" begin
     p = CFTime.Period(10, 1)
@@ -28,104 +28,104 @@ end
     @test one(p) == CFTime.Period(1, 1)
 end
 
-tuf = (2,3,4,5,6,7,8) # day, hour, minute, seconds, ...
+tuf = (2, 3, 4, 5, 6, 7, 8) # day, hour, minute, seconds, ...
 factor = 1
 exponent = -9
-p = CFTime.Period(tuf,factor,exponent)
+p = CFTime.Period(tuf, factor, exponent)
 @test timetuplefrac(p)[1:length(tuf)] == tuf
 
 @test CFTime._type(p) == CFTime._type(typeof(p))
 
-@test one(p) == CFTime.Period(1,factor,exponent)
+@test one(p) == CFTime.Period(1, factor, exponent)
 # test promotion rules
 
-p1 = CFTime.Period(1,:second)
-p2 = CFTime.Period(1000,:millisecond)
+p1 = CFTime.Period(1, :second)
+p2 = CFTime.Period(1000, :millisecond)
 
-@test promote_type(typeof(p1),typeof(p2)) == typeof(p2)
+@test promote_type(typeof(p1), typeof(p2)) == typeof(p2)
 
-pp1,pp2 = promote(p1,p2)
+pp1, pp2 = promote(p1, p2)
 @test pp1 === pp2
 
 
-pp1,pp2 = promote(p1,Dates.Day(1))
+pp1, pp2 = promote(p1, Dates.Day(1))
 @test pp1 == p1
 @test pp2 == Dates.Day(1)
 @test typeof(pp2) == typeof(pp1)
 
 
-pp1,pp2 = promote(p1,Dates.Hour(1))
+pp1, pp2 = promote(p1, Dates.Hour(1))
 @test pp1 == p1
 @test pp2 == Dates.Hour(1)
 @test typeof(pp2) == typeof(pp1)
 
 # missing
 
-p = CFTime.Period(1,:second)
+p = CFTime.Period(1, :second)
 @test ismissing(p == missing)
 @test ismissing(missing == p)
-@test ismissing(DateTimeStandard(2000,1,1) + missing)
-@test ismissing(missing + DateTimeStandard(2000,1,1))
+@test ismissing(DateTimeStandard(2000, 1, 1) + missing)
[…]
