+++ src/conversions.jl
@@ -1,24 +1,24 @@
-function datenum(::Type{T}, y, m, d) where T <: AbstractCFDateTime
+function datenum(::Type{T}, y, m, d) where {T <: AbstractCFDateTime}
     cm = _cum_month_length(T)
     # turn year equal to -1 (1 BC) into year = 0
     if (y < 0) && !_hasyear0(T)
-        y = y+1
+        y = y + 1
     end
 
     if m < 1 || m > 12
         error("invalid month $(m)")
     end
 
-    if d < 1 || d > (cm[m+1] - cm[m])
-        error("invalid day $(d) in $(@sprintf("%04d-%02d-%02d",y,m,d))")
+    if d < 1 || d > (cm[m + 1] - cm[m])
+        error("invalid day $(d) in $(@sprintf("%04d-%02d-%02d", y, m, d))")
     end
 
-    return cm[end] * (y-1) + cm[m] + (d-1)
+    return cm[end] * (y - 1) + cm[m] + (d - 1)
 end
 
 # Calendar with regular month-length
 
-function findmonth(cm,t2)
+function findmonth(cm, t2)
     mo = length(cm)
     while cm[mo] > t2
         mo -= 1
@@ -26,24 +26,24 @@ function findmonth(cm,t2)
     return mo
 end
 
-@inline function datetuple_ymd(::Type{T},timed_::Number) where T <: AbstractCFDateTime
+@inline function datetuple_ymd(::Type{T}, timed_::Number) where {T <: AbstractCFDateTime}
     cm = _cum_month_length(T)
     y = fld(Int64(timed_), cm[end])
-    t2 = Int64(timed_) - cm[end]*y
+    t2 = Int64(timed_) - cm[end] * y
 
     # find month
-    mo = findmonth(cm,t2)
+    mo = findmonth(cm, t2)
     d = t2 - cm[mo]
 
     # day and year start at 1 (not zero)
-    d = d+1
-    y = y+1
+    d = d + 1
+    y = y + 1
 
     if (y <= 0) && !_hasyear0(T)
-        y = y-1
+        y = y - 1
     end
 
-    return (y,mo,d)
+    return (y, mo, d)
 end
 
 @inline _cum_month_length(::Type{DateTimeAllLeap}) =
@@ -56,7 +56,7 @@ end
     (0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330, 360)
 
 
-function validargs(::Type{T},arg...) where T <: AbstractCFDateTime
+function validargs(::Type{T}, arg...) where {T <: AbstractCFDateTime}
     return nothing
 end
 
@@ -83,31 +83,33 @@ Dates.Day(dt - reinterpret(DateTimeStandard,dt))
 ```
 
 """
-function reinterpret(::Type{T1}, dt::DateTime) where T1 <: AbstractCFDateTime
-   return T1(
-       Dates.year(dt),Dates.month(dt),Dates.day(dt),
-       Dates.hour(dt),Dates.minute(dt),Dates.second(dt),
-       Dates.millisecond(dt))
+function reinterpret(::Type{T1}, dt::DateTime) where {T1 <: AbstractCFDateTime}
+    return T1(
+        Dates.year(dt), Dates.month(dt), Dates.day(dt),
+        Dates.hour(dt), Dates.minute(dt), Dates.second(dt),
+        Dates.millisecond(dt)
+    )
 end
 
-function reinterpret(::Type{DateTime}, dt::T2) where T2 <: AbstractCFDateTime
-   return DateTime(
-       Dates.year(dt),Dates.month(dt),Dates.day(dt),
-       Dates.hour(dt),Dates.minute(dt),Dates.second(dt),
[…]
