### 2025-08-26 7b9dff1
format with runic (examples and docs/make.jl)
