### 2025-08-26 JuliaGeo/CFTime.jl#37 @Alexander-Barth
I added the formatting tool runic, in these commits:

7b216b03fff8fd00eaf95e0ee9f197116db4fd70
7b9dff10f38b46e5ff6a6d71c2697ab05c131883
2a51d3a5df5b20b0d4ff57f3b22f01b76108aa80
758ed04f279c431f435b748962247478f6a7d36f

(I hope that this will not cause frustration for new developers)
