### 2025-08-26 #8487 @MartinuzziFrancesco
Thanks for checking in! With the last issues now resolved, I’ve been able to tick off all my final boxes for the review. Well done on addressing them thoroughly, I believe this package is a valuable addition to the ecosystem!
