### 2025-08-29 0397454
fix typo in example code
