### 2025-09-04 90730fc
methods lcm, gcd, gcdx, rem, sign, signbit for periods (issue #60)
