+++ src/CFTime.jl
@@ -50,9 +50,12 @@ import Dates:
     len,
     monthday,
     yearmonth,
-    yearmonthday
+    yearmonthday,
+    value
 
-import Base: +, -, *, /, div, isless, string, show, convert, reinterpret, ==,
+import Base: +, -, *, /, div, rem, lcm, gcd, gcdx, isless,
+    abs, sign, signbit,
+    string, show, convert, reinterpret, ==,
     promote_rule, floor, mod, round, typemax, typemin
 
 include("constants.jl")
