+++ src/period.jl
@@ -24,6 +24,8 @@ _type(::Type{Period{T, Tfactor, Texponent}}) where {T, Tfactor, Texponent} = T
 _factor(::Type{Period{T, Tfactor, Texponent}}) where {T, Tfactor, Texponent} = unwrap(Tfactor)
 _exponent(::Type{Period{T, Tfactor, Texponent}}) where {T, Tfactor, Texponent} = unwrap(Texponent)
 
+Dates.value(p::Period) = p.duration
+
 function Base.zero(p::Period{T, Tfactor, Texponent}) where {T, Tfactor, Texponent}
     return Period{T, Tfactor, Texponent}(0)
 end
@@ -33,11 +35,11 @@ function Base.one(p::Period{T, Tfactor, Texponent}) where {T, Tfactor, Texponent
 end
 
 function Base.abs(p::Period{T, Tfactor, Texponent}) where {T, Tfactor, Texponent}
-    return Period{T, Tfactor, Texponent}(abs(Dates.value(p)))
+    return Period{T, Tfactor, Texponent}(abs(value(p)))
 end
 
-
-Dates.value(p::Period) = p.duration
+Base.sign(x::Period) = sign(value(x))
+Base.signbit(x::Period) = signbit(value(x))
 
 # helper functions for _timetuple
 @inline __tf(result, time) = result
@@ -228,8 +230,8 @@ end
 
 # operations between two CFTime.Periods (or Dates.Period)
 
-#   operators returning a CFTime.Period
-for op in (:+, :-, :mod)
+# operators returning a CFTime.Period
+for op in (:+, :-, :mod, :lcm, :gcd, :rem)
     @eval begin
         function $op(p1::Period{T, Tfactor, Texponent}, p2::Period{T, Tfactor, Texponent}) where {T, Tfactor, Texponent}
             return Period{T, Tfactor, Texponent}($op(p1.duration, p2.duration))
@@ -237,7 +239,12 @@ for op in (:+, :-, :mod)
     end
 end
 
-#   divisions and boolean functions
+function Base.gcdx(a::T, b::T) where {T <: Period}
+    (g, x, y) = gcdx(value(a), value(b))
+    return (T(g), x, y)
+end
+
+# operators not returning a CFTime.Period
 for op in (:/, :div, :(==), :isless)
     @eval begin
         function $op(p1::Period{T, Tfactor, Texponent}, p2::Period{T, Tfactor, Texponent}) where {T, Tfactor, Texponent}
@@ -247,7 +254,7 @@ for op in (:/, :div, :(==), :isless)
 end
 
 
-for op in (:+, :-, :/, :div, :mod, :(==), :isless)
+for op in (:+, :-, :/, :div, :mod, :(==), :isless, :lcm, :gcd, :gcdx, :rem)
     @eval begin
         $op(p1::Period, p2::Period) = $op(promote(p1, p2)...)
         $op(p1::Period, p2::Dates.Period) = $op(promote(p1, p2)...)
@@ -264,6 +271,7 @@ for op in (:*, :/, :div)
         end
     end
 end
+
 *(v::Number, p::Period) = p * v
 
 function -(p::Period{T, Tfactor, Texponent}) where {T, Tfactor, Texponent}
