### 2025-09-17 JuliaGeo/CFTime.jl#61 @mengqi-z
JOSS Paper Proofread
Hi @Alexander-Barth,

I have a few minor comments that should be addressed in the JOSS paper: openjournals/joss-reviews#8487.

1. Line 11: CF is already defined in line 6, so it does not need to define again.
2. Line 20: Software or function names (e.g., CFTime, CFTime.jl, ClimaOcean.jl) should be formatted in code style (e.g., \`CFTime.jl\`) in `paper.md`. Please update all software names consistently throughout the paper.
3. In `paper.bib`, there appear to be several references that are not cited in `paper.md`. Please remove any unused references.

Thanks!
