### 2025-09-18 1be7a57
remove unused reference (issue #61)
