+++ paper/paper.bib
@@ -8,15 +8,6 @@
   doi = "10.62973/21-050r1"
 }
 
-@article{Barth2022,
-  author = {A. Barth and A. Alvera-Azc{\'{a}}rate and C. Troupin and J.-M. Beckers},
-  title = {{DINCAE} 2.0: multivariate convolutional neural network with error estimates to reconstruct sea surface temperature satellite and altimetry observations},
-  year = 2022,
-  publisher = {Copernicus {GmbH}},
-  journal = {Geoscientific Model Development},
-  doi = {10.5194/gmd-2021-353},
-}
-
 @article{OceananigansJOSS,
   author = {Ali Ramadhan and Gregory LeClaire Wagner and Chris Hill and Jean-Michel Campin and Valentin Churavy and Tim Besard and Andre Souza and Alan Edelman and Raffaele Ferrari and John Marshall},
   title = {{Oceananigans.jl: Fast and friendly geophysical fluid dynamics on GPUs}},
@@ -81,32 +72,12 @@
   url = {https://cfconventions.org/Data/cf-conventions/cf-conventions-1.0/build/cf-conventions.html}
 }
 
-@Misc{COARDS,
-  title =     {COARDS NetCDF Conventions},
-  month =     {Feb},
-  year =      {1995},
-  url =    {https://ferret.pmel.noaa.gov/Ferret/documentation/coards-netcdf-conventions}
-}
-
 @manual{Octave,
   title     = {{GNU Octave} version 10.1.0 manual: a high-level interactive language for numerical computations},
   author    = {John W. Eaton and David Bateman and S{\o}ren Hauberg and Rik Wehbring},
   year      = {2025},
   url       = {https://www.gnu.org/software/octave/doc/v10.1.0/},
 }
-
-@article{Hassell2017,
-  AUTHOR = {Hassell, D. and Gregory, J. and Blower, J. and Lawrence, B. N. and Taylor, K. E.},
-  TITLE = {A data model of the Climate and Forecast metadata conventions (CF-1.6) with a software implementation (cf-python v2.1)},
-  JOURNAL = {Geoscientific Model Development},
-  VOLUME = {10},
-  YEAR = {2017},
-  NUMBER = {12},
-  PAGES = {4619--4646},
-  URL = {https://gmd.copernicus.org/articles/10/4619/2017/},
-  DOI = {10.5194/gmd-10-4619-2017}
-}
-
 
 @techreport{SeaDataNet_format,
   title = "SeaDataNet. Datafile formats. ODV, MEDATLAS, NETCDF",
@@ -118,16 +89,6 @@
   editor = "SeaDataNet"
 }
 
-
-@InProceedings{mckinney-proc-scipy-2010,
-  author    = { {W}es {M}c{K}inney },
-  title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
-  booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
-  pages     = { 56 - 61 },
-  year      = { 2010 },
-  editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
-  doi       = { 10.25080/Majora-92bf1922-00a }
-}
 
 @Article{vanVerseveld2024,
   AUTHOR = {van Verseveld, W. J. and Weerts, A. H. and Visser, M. and Buitink, J. and Imhoff, R. O. and Boisgontier, H. and Bouaziz, L. and Eilander, D. and Hegnauer, M. and ten Velden, C. and Russell, B.},
@@ -211,45 +172,6 @@
 
 
 
-@software{Whitaker2024,
-  author       = {Jeff Whitaker and
-                  Spencer Clark and
-                  Constantine Khrulev and
-                  Filipe and
[…]
