### 2025-09-18 8affcd6
proofreading from Editor (issue #61)
