+++ paper/paper.md
@@ -22,35 +22,35 @@ bibliography: paper.bib
 # Summary
 
 
-Climate and Forecast (CF) conventions are a metadata standard for Earth data [@Eaton2024] and are mainly used in oceanography and meteorology.
-The CF conventions were originally proposed for the NetCDF storage format, but they are also increasingly used with other formats like Zarr [@OGC_Zarr] and GRIB ([GRIBDatasets](https://github.com/JuliaGeo/GRIBDatasets.jl)).
+The Climate and Forecast (CF) conventions are a metadata standard for Earth data [@Eaton2024] and are mainly used in oceanography and meteorology.
+The CF conventions were originally proposed for the NetCDF storage format, but they are also increasingly used with other formats like Zarr [@OGC_Zarr] and GRIB ([`GRIBDatasets.jl`](https://github.com/JuliaGeo/GRIBDatasets.jl)).
 
-Since the initial release of the Climate and Forecast (CF) conventions [@Eaton2003], the encoding of time instances has been standardized. The Julia package CFTime implements various calendars that have been standardized in the frame of these conventions. It also supports some arithmetic operations for example, computing the duration between two time instances, ordering time instances and creating time ranges. The time origin and resolution are flexible ranging from days to attoseconds.
+Since the initial release of the CF conventions [@Eaton2003], the encoding of time instances has been standardized. The Julia package `CFTime` implements various calendars that have been standardized in the frame of these conventions. It also supports some arithmetic operations for example, computing the duration between two time instances, ordering time instances and creating time ranges. The time origin and resolution are flexible ranging from days to attoseconds.
 
 
 # Statement of need
 
 In many Earth science disciplines and beyond, expressing a time instance and a duration is essential. The CF conventions provide a rich and flexible
-framework for handling time, equally applicable to observations and model data. To our knowledge, CFTime.jl is the only package in the Julia ecosystem that implements the time structures standardized by the CF conventions. While almost all datasets used in Earth Science use dates after the year 1582, some datasets or software systems use a time origin before this date, which makes it necessary to handle the transition from Julian to Gregorian calendar [@Octave; @SeaDataNet_format].
-Some users also expressed the need for microseconds and nanoseconds as time resolution even if they are rarely used in typical Earth science applications (CFTime github [issue 18](https://github.com/JuliaGeo/CFTime.jl/issues/18) and NCDataset github [issue 248](https://github.com/JuliaGeo/NCDatasets.jl/issues/248)).
+framework for handling time, equally applicable to observations and model data. To our knowledge, `CFTime.jl` is the only package in the Julia ecosystem that implements the time structures standardized by the CF conventions. While almost all datasets used in Earth science use dates after the year 1582, some datasets or software systems  [e.g. @Octave; @SeaDataNet_format] use a time origin before this date, which makes it necessary to handle the transition from the Julian to the Gregorian calendar.
+Some users also expressed the need for microseconds and nanoseconds as time resolutions even if they are rarely used in typical Earth science applications (`CFTime` github [issue 18](https://github.com/JuliaGeo/CFTime.jl/issues/18) and `NCDatasets` github [issue 248](https://github.com/JuliaGeo/NCDatasets.jl/issues/248)).
 
[…]
