### 2025-09-18 a95c910
polishing the text (issue #61)
