### 2025-09-18 1e30e55
update number of dependencies (issue #61)
