+++ paper/paper.md
@@ -34,7 +34,7 @@ In many Earth science disciplines and beyond, expressing a time instance and a d
 framework for handling time, equally applicable to observations and model data. To our knowledge, `CFTime.jl` is the only package in the Julia ecosystem that implements the time structures standardized by the CF conventions. While almost all datasets used in Earth science use dates after the year 1582, some datasets or software systems  [e.g. @Octave; @SeaDataNet_format] use a time origin before this date, which makes it necessary to handle the transition from the Julian to the Gregorian calendar.
 Some users also expressed the need for microseconds and nanoseconds as time resolutions even if they are rarely used in typical Earth science applications (`CFTime` github [issue 18](https://github.com/JuliaGeo/CFTime.jl/issues/18) and `NCDatasets` github [issue 248](https://github.com/JuliaGeo/NCDatasets.jl/issues/248)).
 
-As of 31 March 2025, 119 Julia packages depend directly or indirectly on `CFTime` (excluding optional dependencies). For example, `CFTime` is used by numerical models, such as `ClimaOcean.jl`, a framework for realistic ocean and coupled sea-ice simulations based on `Oceananigans.jl` [@OceananigansJOSS], the hydrological modeling package `Wflow.jl` [@vanVerseveld2024] and `AIBECS.jl`, a modeling framework for global marine biogeochemical cycles [@Pasquier2022].
+As of 18 September 2025, 132 Julia packages depend directly or indirectly on `CFTime` (excluding optional dependencies). For example, `CFTime` is used by numerical models, such as `ClimaOcean.jl`, a framework for realistic ocean and coupled sea-ice simulations based on `Oceananigans.jl` [@OceananigansJOSS], the hydrological modeling package `Wflow.jl` [@vanVerseveld2024] and `AIBECS.jl`, a modeling framework for global marine biogeochemical cycles [@Pasquier2022].
 
 Several data-related packages also make direct or indirect use of `CFTime`, such as the NetCDF manipulation package `NCDatasets.jl` [@Barth2024], the gridded data processing package `YAXArrays.jl` [@Gans2023] and packages for handling in-situ data from various observing platforms (`OceanRobots.jl` [@Forget2024] and `ArgoData.jl` [@Forget2025]).
 
