### 2025-09-29 #8487 @mengqi-z
Thanks @Alexander-Barth ! Could you also update the current version number in your Zenodo archive (`Version v1`) to align with the release version number (`Version v0.2.4`)?
