### 2025-09-30 #8487 @Alexander-Barth
Thanks a lot @mengqi-z for spitting this inconsistency! I just updated the version number in Zenodo.
