### 2025-09-30 #8487 @kthyng
Congratulations on your new publication @Alexander-Barth! Many thanks to editor @mengqi-z and to reviewers @haakon-e, @alextbradley, and @MartinuzziFrancesco for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
